Fixed bug with payment not showing correctly as double

+++ backend/pos-core-api/ORM/TransactionORM.cs
@@ -273,7 +273,7 @@ namespace pos_core_api.ORM
                         Method = itemReader.IsDBNull("method") ? "" : itemReader.GetString("method"),
                         PayId = itemReader.IsDBNull("payid") ? 0 : itemReader.GetUInt32("payid"),
                         Number = itemReader.IsDBNull("Number") ? "" : itemReader.GetString("Number"),
-                        Amount = itemReader.IsDBNull("Amount") ? 0 : itemReader.GetUInt32("Amount")
+                        Amount = itemReader.IsDBNull("Amount") ? 0 : itemReader.GetDouble("Amount")
                     });
                 }
             }
